requirements.txt
alphasquared-py==0.3.0
backoff==2.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coinbase-advanced-py==1.8.2
coinbase-advancedtrade-python==0.3.3
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pymongo==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.10.0
sortedcontainers==2.4.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
websocket_client==0.40.0
websockets==13.1
Werkzeug==3.1.3
